Move bulk reader and indexer to separate module
