+++ examples/load.rs
@@ -5,54 +5,12 @@ extern crate log;
 
 extern crate error_chain;
 
-use std::path::PathBuf;
-use std::sync::{
-    mpsc::{Receiver, SyncSender}, Arc, Mutex,
-};
-
 use electrs::{
-    bulk::{Parser, FINISH_MARKER}, config::Config, daemon::Daemon, errors::*, metrics::Metrics,
-    store::{DBStore, Row, StoreOptions, WriteStore}, util::{spawn_thread, SyncChannel},
+    bulk, config::Config, daemon::Daemon, errors::*, metrics::Metrics,
+    store::{DBStore, StoreOptions},
 };
 
 use error_chain::ChainedError;
-
-type JoinHandle = std::thread::JoinHandle<Result<()>>;
-type BlobReceiver = Arc<Mutex<Receiver<(Vec<u8>, PathBuf)>>>;
-
-fn start_reader(blk_files: Vec<PathBuf>, parser: Arc<Parser>) -> (BlobReceiver, JoinHandle) {
-    let chan = SyncChannel::new(0);
-    let blobs = chan.sender();
-    let handle = spawn_thread("bulk_read", move || -> Result<()> {
-        for path in blk_files {
-            blobs
-                .send((parser.read_blkfile(&path)?, path))
-                .expect("failed to send blk*.dat contents");
-        }
-        Ok(())
-    });
-    (Arc::new(Mutex::new(chan.into_receiver())), handle)
-}
-
-fn start_indexer(
-    blobs: BlobReceiver,
-    parser: Arc<Parser>,
-    rows: SyncSender<(Vec<Row>, PathBuf)>,
-) -> JoinHandle {
-    spawn_thread("bulk_index", move || -> Result<()> {
-        loop {
-            let msg = blobs.lock().unwrap().recv();
-            if let Ok((blob, path)) = msg {
-                rows.send((parser.index_blkfile(blob)?, path))
-                    .expect("failed to send indexed rows")
-            } else {
-                debug!("no more blocks to index");
-                break;
-            }
-        }
-        Ok(())
-    })
-}
 
 fn run(config: Config) -> Result<()> {
     if config.db_path.exists() {
@@ -69,37 +27,8 @@ fn run(config: Config) -> Result<()> {
         config.network_type,
         &metrics,
     )?;
-    bulk::set_open_files_limit(2048); // twice the default `ulimit -n` value
     let store = DBStore::open(&config.db_path, StoreOptions { bulk_import: true });
-    let blk_files = daemon.list_blk_files()?;
-    let parser = Arc::new(Parser::new(&daemon, &metrics)?);
-    let (blobs, reader) = start_reader(blk_files, parser.clone());
-    let rows_chan = SyncChannel::new(0);
-    let indexers: Vec<JoinHandle> = (0..2)
-        .map(|_| start_indexer(blobs.clone(), parser.clone(), rows_chan.sender()))
-        .collect();
-    spawn_thread("bulk_writer", move || {
-        for (rows, path) in rows_chan.into_receiver() {
-            trace!("indexed {:?}: {} rows", path, rows.len());
-            store.write(rows);
-        }
-        reader
-            .join()
-            .expect("reader panicked")
-            .expect("reader failed");
-
-        indexers.into_iter().for_each(|i| {
-            i.join()
-                .expect("indexer panicked")
-                .expect("indexing failed")
-        });
-        store.write(vec![parser.last_indexed_row()]);
-        store.flush();
-        store.compact(); // will take a while.
-        store.put(FINISH_MARKER, b"");
-    }).join()
-        .expect("writer panicked");
-    Ok(())
+    bulk::index(&daemon, &metrics, store)
 }
 
 fn main() {
