+++ src/bulk.rs
@@ -7,20 +7,23 @@ use libc;
 use std::collections::HashSet;
 use std::fs;
 use std::io::{Cursor, Seek, SeekFrom};
-use std::path::Path;
-use std::sync::Mutex;
+use std::path::{Path, PathBuf};
+use std::sync::{
+    mpsc::{Receiver, SyncSender}, Arc, Mutex,
+};
+use std::thread;
 
 use daemon::Daemon;
 use index::{index_block, last_indexed_block};
 use metrics::{CounterVec, Histogram, HistogramOpts, HistogramVec, MetricOpts, Metrics};
-use store::Row;
-use util::HeaderList;
+use store::{DBStore, Row, WriteStore};
+use util::{spawn_thread, HeaderList, SyncChannel};
 
 use errors::*;
 
-pub const FINISH_MARKER: &'static [u8] = b"F";
+const FINISH_MARKER: &'static [u8] = b"F";
 
-pub struct Parser {
+struct Parser {
     magic: u32,
     current_headers: HeaderList,
     indexed_blockhashes: Mutex<HashSet<Sha256dHash>>,
@@ -31,8 +34,8 @@ pub struct Parser {
 }
 
 impl Parser {
-    pub fn new(daemon: &Daemon, metrics: &Metrics) -> Result<Parser> {
-        Ok(Parser {
+    fn new(daemon: &Daemon, metrics: &Metrics) -> Result<Arc<Parser>> {
+        Ok(Arc::new(Parser {
             magic: daemon.magic(),
             current_headers: load_headers(daemon)?,
             indexed_blockhashes: Mutex::new(HashSet::new()),
@@ -49,10 +52,10 @@ impl Parser {
                 "parse_bytes_read",
                 "# of bytes read (from blk*.dat)",
             )),
-        })
+        }))
     }
 
-    pub fn last_indexed_row(&self) -> Row {
+    fn last_indexed_row(&self) -> Row {
         let indexed_blockhashes = self.indexed_blockhashes.lock().unwrap();
         let last_header = self.current_headers
             .iter()
@@ -63,7 +66,7 @@ impl Parser {
         last_indexed_block(last_header.hash())
     }
 
-    pub fn read_blkfile(&self, path: &Path) -> Result<Vec<u8>> {
+    fn read_blkfile(&self, path: &Path) -> Result<Vec<u8>> {
         let timer = self.duration.with_label_values(&["read"]).start_timer();
         let blob = fs::read(&path).chain_err(|| format!("failed to read {:?}", path))?;
         timer.observe_duration();
@@ -71,7 +74,7 @@ impl Parser {
         return Ok(blob);
     }
 
-    pub fn index_blkfile(&self, blob: Vec<u8>) -> Result<Vec<Row>> {
+    fn index_blkfile(&self, blob: Vec<u8>) -> Result<Vec<Row>> {
         let timer = self.duration.with_label_values(&["parse"]).start_timer();
         let blocks = parse_blocks(blob, self.magic)?;
         timer.observe_duration();
@@ -148,16 +151,93 @@ fn load_headers(daemon: &Daemon) -> Result<HeaderList> {
     Ok(headers)
 }
 
-pub fn set_open_files_limit(limit: u64) {
+fn set_open_files_limit(limit: u64) {
     let resource = libc::RLIMIT_NOFILE;
-    let mut rlim = libc::rlimit { rlim_cur: 0, rlim_max: 0 };
+    let mut rlim = libc::rlimit {
+        rlim_cur: 0,
+        rlim_max: 0,
+    };
     let result = unsafe { libc::getrlimit(resource, &mut rlim) };
     if result < 0 {
         panic!("getrlimit() failed: {}", result);
     }
-    rlim.rlim_cur = limit;  // set softs limit only.
+    rlim.rlim_cur = limit; // set softs limit only.
     let result = unsafe { libc::setrlimit(resource, &rlim) };
     if result < 0 {
         panic!("setrlimit() failed: {}", result);
     }
+}
+
+type JoinHandle = thread::JoinHandle<Result<()>>;
+type BlobReceiver = Arc<Mutex<Receiver<(Vec<u8>, PathBuf)>>>;
+
+fn start_reader(blk_files: Vec<PathBuf>, parser: Arc<Parser>) -> (BlobReceiver, JoinHandle) {
+    let chan = SyncChannel::new(0);
+    let blobs = chan.sender();
+    let handle = spawn_thread("bulk_read", move || -> Result<()> {
+        for path in blk_files {
+            blobs
+                .send((parser.read_blkfile(&path)?, path))
+                .expect("failed to send blk*.dat contents");
+        }
+        Ok(())
+    });
+    (Arc::new(Mutex::new(chan.into_receiver())), handle)
+}
+
+fn start_indexer(
+    blobs: BlobReceiver,
+    parser: Arc<Parser>,
+    writer: SyncSender<(Vec<Row>, PathBuf)>,
+) -> JoinHandle {
+    spawn_thread("bulk_index", move || -> Result<()> {
+        loop {
+            let msg = blobs.lock().unwrap().recv();
+            if let Ok((blob, path)) = msg {
+                let rows = parser
+                    .index_blkfile(blob)
+                    .chain_err(|| format!("failed to index {:?}", path))?;
+                writer
+                    .send((rows, path))
+                    .expect("failed to send indexed rows")
+            } else {
+                debug!("no more blocks to index");
+                break;
+            }
+        }
+        Ok(())
+    })
+}
+
+pub fn index(daemon: &Daemon, metrics: &Metrics, store: DBStore) -> Result<()> {
+    set_open_files_limit(2048); // twice the default `ulimit -n` value
+    let blk_files = daemon.list_blk_files()?;
+    let parser = Parser::new(daemon, metrics)?;
+    let (blobs, reader) = start_reader(blk_files, parser.clone());
+    let rows_chan = SyncChannel::new(0);
+    let indexers: Vec<JoinHandle> = (0..2)
+        .map(|_| start_indexer(blobs.clone(), parser.clone(), rows_chan.sender()))
+        .collect();
+    spawn_thread("bulk_writer", move || {
+        for (rows, path) in rows_chan.into_receiver() {
+            trace!("indexed {:?}: {} rows", path, rows.len());
+            store.write(rows);
+        }
+        reader
+            .join()
+            .expect("reader panicked")
+            .expect("reader failed");
+
+        indexers.into_iter().for_each(|i| {
+            i.join()
+                .expect("indexer panicked")
+                .expect("indexing failed")
+        });
+        store.write(vec![parser.last_indexed_row()]);
+        store.flush();
+        store.compact(); // will take a while.
+        store.put(FINISH_MARKER, b"");
+    }).join()
+        .expect("writer panicked");
+    Ok(())
 }
